Refactor component locations
