+++ src/components/bugout/YourColorChosenModal.js
